requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio>=3.4.3
aiofiles>=23.0.0
typing-extensions>=4.0.0
rich>=13.0.0
aiosqlite>=0.19.0
